llm_project/requirements.txt
openai
google-generativeai
transformers
torch
pyyaml
